Safari-609.1.11.1/LayoutTests/platform/wpe/imported/w3c/web-platform-tests/xhr/xmlhttprequest-network-error-sync-expected.txt

CONSOLE MESSAGE: line 23: Load cannot follow more than 20 redirections
CONSOLE MESSAGE: line 23: XMLHttpRequest cannot load http://localhost:8800/xhr/resources/infinite-redirects.py?page=default&type=302&mix=0 due to access control checks.

PASS XMLHttpRequest: members during network errors (sync) 
